Se agrego la clase jugador, y ahora alos objetos que pueden ser propios de un jugador se le agrego un id que hace referencia a ese jugador, solo el jugador que el mismo ID que el objeto puede seleccionar ese objeto. Se agrego un edificio para la creacion de unidades, este tambien tiene la ID del jugador que lo puede usar y genera unidades con el ID del jugador alrededor de el

+++ geEngine-RTS/GameRTS/Include/RTSFunctionsCoords.h
@@ -108,8 +108,8 @@ namespace COORDS {
       //tileCoord.x += 1;
       //pixCoord.x = (tileCoord.x - tileCoord.y) << GameOptions::BITSFHT_TILEHALFSIZE.x;
       //pixCoord.y = (tileCoord.x + tileCoord.y) << GameOptions::BITSFHT_TILEHALFSIZE.y;
-      pixCoord.x = pixX ;
-      pixCoord.y = pixY ;
+      pixCoord.x = (float)pixX ;
+      pixCoord.y = (float)pixY ;
 
       scrX = pixCoord.x + PreCalc_ScreenDeface.x;
       scrY = pixCoord.y + PreCalc_ScreenDeface.y;
